third_party/blink/web_tests/platform/generic/fast/events/popup-blocked-from-wrong-event-expected.txt

click here 

Tests that the plugin container doesn't generate user gestures

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.open('about:blank') is defined.
PASS window.open('about:blank') is null
PASS successfullyParsed is true

TEST COMPLETE
